+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/Episode.swift
@@ -15,7 +15,6 @@ public class Episode: NSObject, BaseEpisode {
     @objc public var episodeDescription: String?
     @objc public var episodeStatus = 0 as Int32
     @objc public var fileType: String?
-    @objc public var contentType: String?
     @objc public var keepEpisode = false
     @objc public var playedUpTo: Double = 0
     @objc public var duration: Double = 0
@@ -46,10 +45,14 @@ public class Episode: NSObject, BaseEpisode {
     @objc public var hasOnlyUuid = false
     @objc public var deselectedChapters: String?
     @objc public var deselectedChaptersModified = 0 as Int64
-    @objc public var rawMetadata: String?
 
     public var hasBookmarks: Bool {
-        DataManager.sharedManager.bookmarks.bookmarkCount(forEpisode: uuid) > 0
+        // This wil cause a regression in which the bookmarks won't be displayed
+        // for episodes with bookmarks.
+        // However, this call is happening on the main thread and can block the whole app.
+        // We will re-add this again in a way that's not a blocker
+        //DataManager.sharedManager.bookmarks.bookmarkCount(forEpisode: uuid) > 0
+        false
     }
 
     public var isUserEpisode: Bool {
@@ -178,22 +181,7 @@ public class Episode: NSObject, BaseEpisode {
         taggableId()
     }
 
-    private var parsedMetadata: Episode.Metadata? = nil
-
-    public var metadata: Episode.Metadata? {
-        if let parsedMetadata {
-            return parsedMetadata
-        } else if let data = rawMetadata?.data(using: .utf8) {
-            // Decode and cache metadata
-            let decoder = JSONDecoder()
-            decoder.keyDecodingStrategy = .convertFromSnakeCase
-            let metadata = try? decoder.decode(Episode.Metadata.self, from: data)
-            parsedMetadata = metadata
-            return metadata
-        }
-
-        return nil
-    }
+    // MARK: - Metadata
 
     public struct Metadata: Decodable {
         public let showNotes: String?