+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -10,7 +10,7 @@ public enum FeatureFlag: String, CaseIterable {
     /// Whether End Of Year feature is enabled
     case endOfYear
 
-    /// Enable the new show notes endpoint plus embedded episode artwork
+    /// Enable show notes using the new endpoint
     case newShowNotesEndpoint
 
     /// Enable retrieving episode artwork from the RSS feed
@@ -50,6 +50,10 @@ public enum FeatureFlag: String, CaseIterable {
     /// show UpNext tab on the main tab bar
     case upNextOnTabBar
 
+    /// When enabled it updates the code on filter callback to use a safer method to convert unmanaged player references
+    /// This is to fix this: https://a8c.sentry.io/share/issue/39a6d2958b674ec3b7a4d9248b4b5ffa/
+    case defaultPlayerFilterCallbackFix
+
     case newSharing
 
     public var enabled: Bool {
@@ -69,11 +73,11 @@ public enum FeatureFlag: String, CaseIterable {
         case .endOfYear:
             false
         case .newShowNotesEndpoint:
-            true
+            false
         case .episodeFeedArtwork:
-            true // To be enabled, newShowNotesEndpoint needs to be too
+            false
         case .rssChapters:
-            true // To be enabled, newShowNotesEndpoint needs to be too
+            false
         case .newPlayerTransition:
             true
         case .errorLogoutHandling:
@@ -94,16 +98,17 @@ public enum FeatureFlag: String, CaseIterable {
             true
         case .categoriesRedesign:
             true
-        case .upNextOnTabBar:
+        case .defaultPlayerFilterCallbackFix:
             true
+        case .upNextOnTabBar:
+            false
         case .newSharing:
             false
         }
     }
 
     private var shouldEnableSyncedSettings: Bool {
-        // Enabled only out of appstore until we verify that this feature is ready for production.
-        BuildEnvironment.current != .appStore
+        false
     }
 
     /// Remote Feature Flag
@@ -120,6 +125,16 @@ public enum FeatureFlag: String, CaseIterable {
             shouldEnableSyncedSettings ? "new_settings_storage" : nil
         case .settingsSync:
             shouldEnableSyncedSettings ? "settings_sync" : nil
+        case .newShowNotesEndpoint:
+             "new_show_notes"
+         case .episodeFeedArtwork:
+             "episode_artwork"
+         case .rssChapters:
+             "rss_chapters"
+        case .categoriesRedesign:
+            "categories_redesign"
+        case .defaultPlayerFilterCallbackFix:
+            "default_player_filter_callback_fix"
         default:
             nil
         }