+++ podcasts/Analytics/AnalyticsEvent.swift
@@ -287,9 +287,12 @@ enum AnalyticsEvent: String {
 
     case discoverShown
     case discoverCategoryShown
+    case discoverCategoriesPillTapped
     case discoverFeaturedPodcastTapped
     case discoverFeaturedPodcastSubscribed
     case discoverShowAllTapped
+    case discoverCategoryCloseButtonTapped
+    case discoverCategoriesPickerPick
 
     case discoverListImpression
     case discoverListShowAllTapped