+++ podcasts/EpisodePreviewCell.swift
@@ -21,8 +21,6 @@ class EpisodePreviewCell: ThemeableCell {
         episodeTitle.text = episode.title
         if let userEpisode = episode as? UserEpisode {
             episodeImage.setUserEpisode(uuid: userEpisode.uuid, size: .list)
-        } else if let episode = episode as? Episode {
-            episodeImage.setEpisode(episode, size: .list)
         } else {
             episodeImage.setPodcast(uuid: episode.parentIdentifier(), size: .list)
         }