+++ podcasts/GeneralSettingsViewController.swift
@@ -116,7 +116,12 @@ class GeneralSettingsViewController: UIViewController, UITableViewDelegate, UITa
 
             cell.cellLabel.text = L10n.settingsGeneralKeepScreenAwake
 
-            cell.cellSwitch.isOn = Settings.keepScreenAwake
+            if FeatureFlag.newSettingsStorage.enabled {
+                cell.cellSwitch.isOn = SettingsStore.appSettings.keepScreenAwake
+            } else {
+                cell.cellSwitch.isOn = UserDefaults.standard.bool(forKey: Constants.UserDefaults.keepScreenOnWhilePlaying)
+            }
+
             cell.cellSwitch.removeTarget(self, action: nil, for: .valueChanged)
             cell.cellSwitch.addTarget(self, action: #selector(screenLockToggled(_:)), for: .valueChanged)
 
@@ -137,7 +142,11 @@ class GeneralSettingsViewController: UIViewController, UITableViewDelegate, UITa
 
             cell.cellLabel.text = L10n.settingsGeneralAutoOpenPlayer
 
-            cell.cellSwitch.isOn = Settings.openPlayerAutomatically
+            if FeatureFlag.newSettingsStorage.enabled {
+                cell.cellSwitch.isOn = SettingsStore.appSettings.openPlayer
+            } else {
+                cell.cellSwitch.isOn = UserDefaults.standard.bool(forKey: Constants.UserDefaults.openPlayerAutomatically)
+            }
 
             cell.cellSwitch.removeTarget(self, action: nil, for: .valueChanged)
             cell.cellSwitch.addTarget(self, action: #selector(openPlayerToggled(_:)), for: .valueChanged)
@@ -148,7 +157,11 @@ class GeneralSettingsViewController: UIViewController, UITableViewDelegate, UITa
 
             cell.cellLabel.text = L10n.settingsGeneralSmartPlayback
 
-            cell.cellSwitch.isOn = Settings.intelligentResumption
+            if FeatureFlag.newSettingsStorage.enabled {
+                cell.cellSwitch.isOn = SettingsStore.appSettings.intelligentResumption
+            } else {
+                cell.cellSwitch.isOn = UserDefaults.standard.bool(forKey: Constants.UserDefaults.intelligentPlaybackResumption)
+            }
 
             cell.cellSwitch.removeTarget(self, action: nil, for: .valueChanged)
             cell.cellSwitch.addTarget(self, action: #selector(intelligentPlaybackResumptionToggled(_:)), for: .valueChanged)
@@ -429,7 +442,10 @@ class GeneralSettingsViewController: UIViewController, UITableViewDelegate, UITa
     }
 
     @objc private func screenLockToggled(_ sender: UISwitch) {
-        Settings.keepScreenAwake = sender.isOn
+        if FeatureFlag.newSettingsStorage.enabled {
+            SettingsStore.appSettings.keepScreenAwake = sender.isOn
+        }
+        UserDefaults.standard.set(sender.isOn, forKey: Constants.UserDefaults.keepScreenOnWhilePlaying)
         PlaybackManager.shared.updateIdleTimer()
         Settings.trackValueToggled(.settingsGeneralKeepScreenAwakeToggled, enabled: sender.isOn)
     }
@@ -452,12 +468,18 @@ class GeneralSettingsViewController: UIViewController, UITableViewDelegate, UITa
     }
 
     @objc private func openPlayerToggled(_ sender: UISwitch) {
-        Settings.openPlayerAutomatically = sender.isOn
+        if FeatureFlag.newSettingsStorage.enabled {
+            SettingsStore.appSettings.openPlayer = sender.isOn
+        }
+        UserDefaults.standard.set(sender.isOn, forKey: Constants.UserDefaults.openPlayerAutomatically)
         Settings.trackValueToggled(.settingsGeneralOpenPlayerAutomaticallyToggled, enabled: sender.isOn)
     }
 
     @objc private func intelligentPlaybackResumptionToggled(_ sender: UISwitch) {
-        Settings.intelligentResumption = sender.isOn
+        if FeatureFlag.newSettingsStorage.enabled {
+            SettingsStore.appSettings.intelligentResumption = sender.isOn
+        }
+        UserDefaults.standard.set(sender.isOn, forKey: Constants.UserDefaults.intelligentPlaybackResumption)
         Settings.trackValueToggled(.settingsGeneralIntelligentPlaybackToggled, enabled: sender.isOn)
     }