+++ podcasts/MiniPlayerViewController.swift
@@ -208,7 +208,11 @@ class MiniPlayerViewController: SimpleNotificationsViewController {
             setupForEpisode(episode)
             showMiniPlayer()
             let shouldOpenAutomatically: Bool
-            shouldOpenAutomatically = Settings.openPlayerAutomatically
+            if FeatureFlag.newSettingsStorage.enabled {
+                shouldOpenAutomatically = SettingsStore.appSettings.openPlayer
+            } else {
+                shouldOpenAutomatically = UserDefaults.standard.bool(forKey: Constants.UserDefaults.openPlayerAutomatically)
+            }
             if shouldOpenAutomatically || episode.videoPodcast(), lastEpisodeUuidAutoOpened != episode.uuid {
                 lastEpisodeUuidAutoOpened = episode.uuid