+++ podcasts/SinglePodcastViewController.swift
@@ -66,7 +66,7 @@ class SinglePodcastViewController: UIViewController, DiscoverSummaryProtocol {
 
     // MARK: DiscoverSummaryProtocol
 
-    func populateFrom(item: DiscoverItem) {
+    func populateFrom(item: DiscoverItem, region: String?) {
         guard let source = item.source else { return }
 
         self.item = item
@@ -102,7 +102,6 @@ class SinglePodcastViewController: UIViewController, DiscoverSummaryProtocol {
         if let isSponsored = item?.isSponsored, isSponsored {
             typeBadgeLabel.text = L10n.discoverSponsored
             typeBadgeLabel.style = .primaryText02
-            podcastDescription.numberOfLines = 3
         } else {
             typeBadgeLabel.text = L10n.discoverFreshPick
             typeBadgeLabel.style = .support02