+++ podcasts/Strings+Generated.swift
@@ -472,6 +472,8 @@ internal enum L10n {
   internal static var customEpisodeRemoveUpload: String { return L10n.tr("Localizable", "custom_episode_remove_upload") }
   /// Upload to Cloud
   internal static var customEpisodeUpload: String { return L10n.tr("Localizable", "custom_episode_upload") }
+  /// We're moving a few bits and bytes so the app runs faster...
+  internal static var databaseMigration: String { return L10n.tr("Localizable", "database_migration") }
   /// Day listened
   internal static var dayListened: String { return L10n.tr("Localizable", "day_listened") }
   /// Day saved
@@ -1254,6 +1256,8 @@ internal enum L10n {
   internal static var gotIt: String { return L10n.tr("Localizable", "got_it") }
   /// Group Episodes
   internal static var groupEpisodes: String { return L10n.tr("Localizable", "group_episodes") }
+  /// Hang on!
+  internal static var hangOn: String { return L10n.tr("Localizable", "hang_on") }
   /// Clear All
   internal static var historyClearAll: String { return L10n.tr("Localizable", "history_clear_all") }
   /// Clear Listening History
@@ -2814,6 +2818,16 @@ internal enum L10n {
   internal static var sleepTimerCancel: String { return L10n.tr("Localizable", "sleep_timer_cancel") }
   /// End Of Episode
   internal static var sleepTimerEndOfEpisode: String { return L10n.tr("Localizable", "sleep_timer_end_of_episode") }
+  /// In %1$@ episodes
+  internal static func sleepTimerEpisodeCount(_ p1: Any) -> String {
+    return L10n.tr("Localizable", "sleep_timer_episode_count", String(describing: p1))
+  }
+  /// Sleeping in %1$@ episodes
+  internal static func sleepTimerSleepingAfter(_ p1: Any) -> String {
+    return L10n.tr("Localizable", "sleep_timer_sleeping_after", String(describing: p1))
+  }
+  /// Sleeping at the end of the current episode
+  internal static var sleepTimerSleepingAfterCurrentEpisode: String { return L10n.tr("Localizable", "sleep_timer_sleeping_after_current_episode") }
   /// Sleep Timer on, %1$@ remaining
   internal static func sleepTimerTimeRemaining(_ p1: Any) -> String {
     return L10n.tr("Localizable", "sleep_timer_time_remaining", String(describing: p1))